+++ R/predict.R
@@ -16,6 +16,8 @@
 #'   Other options are \code{"confidence"} (for confidence intervals) and
 #'   \code{"prediction"} (for prediction intervals).
 #' @param level Tolerance/confidence level. The default is \code{0.95}.
+#' @param type The prediction type, either on the response scale, link scale (only for
+#'   \code{spglm()} or \code{spgautor()} model objects), or terms scale.
 #' @param local A optional logical or list controlling the big data approximation. If omitted, \code{local}
 #'   is set to \code{TRUE} or \code{FALSE} based on the observed data sample size (i.e., sample size of the fitted
 #'   model object) -- if the sample size exceeds 10,000, \code{local} is
@@ -54,6 +56,8 @@
 #'   If \code{local} is \code{TRUE}, defaults for \code{local} are chosen such
 #'   that \code{local} is transformed into
 #'   \code{list(size = 100, method = "covariance", parallel = FALSE)}.
+#' @param terms If \code{type} is \code{"terms"}, the type of terms to be returned,
+#'   specified via either numeric position or name. The default is all terms are included.
 #' @param ... Other arguments. Only used for models fit using \code{splmRF()}
 #'   or \code{spautorRF()} where \code{...} indicates other
 #'   arguments to \code{ranger::predict.ranger()}.
@@ -103,10 +107,11 @@
 #' predict(spmod, sulfate_preds, interval = "prediction")
 #' augment(spmod, newdata = sulfate_preds, interval = "prediction")
 predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "confidence", "prediction"),
-                         level = 0.95, local, ...) {
+                         level = 0.95, type = c("response", "terms"), local, terms = NULL, ...) {
 
   # match interval argument so the three display
   interval <- match.arg(interval)
+  type <- match.arg(type)
 
   # deal with local
   if (missing(local)) {
@@ -211,6 +216,11 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
   attr(newdata_model, "assign") <- attr_assign[keep_cols]
   attr(newdata_model, "contrasts") <- attr_contrasts
 
+  # call terms if needed
+  if (type == "terms") {
+    return(predict_terms(object, newdata_model, se.fit, interval, level, add_newdata_rows, terms, ...))
+  }
+
   # storing newdata as a list
   newdata_rows_list <- split(newdata, seq_len(NROW(newdata)))
 
@@ -219,7 +229,6 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
 
   # storing newdata as a list
   newdata_list <- mapply(x = newdata_rows_list, y = newdata_model_list, FUN = function(x, y) list(row = x, x0 = y), SIMPLIFY = FALSE)
-
 
   if (interval %in% c("none", "prediction")) {
 
@@ -498,10 +507,11 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
 #' @order 2
 #' @export
 predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none", "confidence", "prediction"),
-                            level = 0.95, local, ...) {
+                            level = 0.95, type = c("response", "terms"), local, terms = NULL, ...) {
 
   # match interval argument so the three display
   interval <- match.arg(interval)
+  type <- match.arg(type)
 
   # deal with local
   if (missing(local)) {
@@ -558,6 +568,13 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
   newdata_model <- newdata_model[, keep_cols, drop = FALSE]
   attr(newdata_model, "assign") <- attr_assign[keep_cols]
   attr(newdata_model, "contrasts") <- attr_contrasts
+
+  # call terms if needed
+  if (type == "terms") {
+    # may want to add add_newdata_rows if we allow spautor prediction for the observed model matrix
+    return(predict_terms(object, newdata_model, se.fit, interval, level, add_newdata_rows = TRUE, terms, ...))
+  }
+
 
   # storing newdata as a list
   newdata_list <- split(newdata, seq_len(NROW(newdata)))